Fix for #6 & some cosmetic changes
Moved get_trade_hist_alias() from bundle.py to api.py, as I believe it makes more sense there. Added a catch for the HTTPError I mentioned in issue #6.

Also added instructions for installing the repo from github, should anyone need it.
